Add separate files for classes NewEventFormComponent and EditEventFormComponent
